requirements.txt
setuptools
wheel
pyyaml
termcolor
requests
